requirements.txt
Flask
Flask-Cors
requests
python-whois
dnspython
google-cloud-storage
gunicorn
celery
redis